Allow the specification of PSO networks
The default choice is a completely connected network, which preserves
the standard PSO functionality. It is implemented by using the global
best value and index as it was already computed, so it should not have
a significant performance impact.

Other choices available for now are a ring and grid topology. These
choices are not currently exposed through the user interface.