Fix detailed models and use table optimization
Validate that the TNNP2006, OVVR, and ORTP models match the F90
implementations as much as possible. In particular, fix several issues
that were the consequence of the lower precision floating point format
used in WebGL shaders.

Additionally, precompute many values that are only depenedent on voltage
and store the results in tables. This optimization seems to result in
speedup between 1.5x and 2x, depending on the model.